requirements.txt
streamlit
beautifulsoup4
PyPDF2
python-docx